매수매도 이벤트시 포트폴리오 코인 구매하는 로직 구현

+++ src/main/java/com/cloneCoin/portfolio/domain/Copy.java
@@ -17,7 +17,7 @@ public class Copy {
     private Long userId;
     private Long leaderId;
     //private Date registerDate;
-    private Long totalInvestAmout;
+    private Double totalInvestAmout;
     //private Long copyProfit;
 
     @ManyToOne
